Downgrade `react-devtools`, `react-devtools-core`
This can be inspected in `package-lock.json` that `react-native` depends on `react-devtools-core@4.19`.

This project, on the other hand, depends on `react-devtools@4.24`. Which case the error `Unsupported Bridge Protocal "0"`, due to the fact that the 2 packages' versions must be the same.

Downgrading the `react-devtools` in the project will fix the issue.